🔨 files structure refacto
folder for hooks, globals, and queries to clean up code and separate
concerns
