+++ front/src/globals/variables.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { IUserInputsRef } from "./Interfaces";
+
+export const GUserInputsRefs: IUserInputsRef = {
+  username: React.createRef(),
+  email: React.createRef(),
+  password: React.createRef(),
+};
