+++ front/src/hooks/UserInfoHooks.tsx
@@ -1,0 +1,20 @@
+import { useState } from "react";
+
+// export const useUsername = () => {
+//   const [showUsername, setShowUsername] = useState(false);
+//   const onClickEditUsername = () => setShowUsername((curent) => !curent);
+//   const hideUsername = () => setShowUsername(false);
+//   return (showUsername);
+// };
+
+// export const [showEmail, setShowEmail] = useState(false);
+// export const onClickEditEmail = () => setShowEmail((curent) => !curent);
+// export const hideEmail = () => setShowEmail(false);
+
+// export const [showPhone, setShowPhone] = useState(false);
+// export const onClickEditPhone = () => setShowPhone((curent) => !curent);
+// export const hidePhone = () => setShowPhone(false);
+
+// export const [showPass, setShowPass] = useState(false);
+// export const onClickEditPass = () => setShowPass((curent) => !curent);
+// export const hidePass = () => setShowPass(false);
