+++ front/src/routes/Auth.tsx
@@ -4,73 +4,15 @@ import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import { Outlet } from "react-router-dom";
 import { useAuth } from "..";
-
-export const userInputsRefs: {
-  username: React.RefObject<HTMLInputElement>;
-  email: React.RefObject<HTMLInputElement>;
-  password: React.RefObject<HTMLInputElement>;
-} = {
-  username: React.createRef(),
-  email: React.createRef(),
-  password: React.createRef(),
-};
-
-let userInfo: {
-  username: string | null;
-  email: string | null;
-  password: string | null;
-  clear: any;
-} = {
-  username: null,
-  email: null,
-  password: null,
-  clear: function () {
-    this.username = null;
-    this.email = null;
-    this.password = null;
-  },
-};
-
-// interface LocationState {
-//   from: {
-//     pathname: string;
-//   };
-// }
-
-const storeUserInfo = (userInfo: any, token: string) => {
-  if (
-    token !==
-      '{"statusCode":403,"message":"Credentials already exist","error":"Forbidden"}' &&
-    token !==
-      '{"statusCode":403,"message":"Invalid Credentials","error":"Forbidden"}'
-  ) {
-    console.log("HERE");
-    const subOne = token.replace('{"access_token":"', "");
-    const subTwo = subOne.replace('"}', "");
-    localStorage.setItem("userToken", subTwo);
-    if (userInfo.username) localStorage.setItem("userName", userInfo.username);
-    else localStorage.setItem("userName", "not-found");
-
-    localStorage.setItem("userEmail", userInfo.email);
-    localStorage.setItem("userPassword", userInfo.password);
-
-    console.log("token: " + localStorage.getItem("userToken"));
-    console.log("userName: " + localStorage.getItem("userName"));
-    console.log("userEmail: " + localStorage.getItem("userEmail"));
-    console.log("userPassword: " + localStorage.getItem("userPassword"));
-  }
-  userInfo.clear();
-};
+import { IUserInfo } from "../globals/Interfaces";
+import { signUp, signIn } from "../queries/authQueries";
+import { GUserInputsRefs } from "../globals/variables";
 
 export default function Auth() {
   let navigate = useNavigate();
-  // let location = useLocation();
   let auth = useAuth(); // subscribe to Auth context
 
   function userSignIn() {
-    // const { from } = (location.state as LocationState) || {
-    //   from: { pathname: "/" },
-    // };
     let email = localStorage.getItem("userEmail");
     if (email)
       auth.signin(email, () => {
@@ -79,59 +21,25 @@ export default function Auth() {
   }
 
   const handleSubmit = (event: any) => {
+    let userInfo: IUserInfo = {
+      username: null,
+      email: null,
+      password: null,
+      clear: function () {
+        this.username = null;
+        this.email = null;
+        this.password = null;
+      },
+    };
+
     event.preventDefault();
-
-    if (userInputsRefs.username.current?.value)
-      userInfo.username = userInputsRefs.username.current.value;
-    userInfo.email = userInputsRefs!.email!.current!.value;
-    userInfo.password = userInputsRefs!.password!.current!.value;
-    if (userInfo.username && userInfo.email && userInfo.password) signUp();
-    else signIn();
-  };
-
-  const signIn = () => {
-    let myHeaders = new Headers();
-    myHeaders.append("Content-Type", "application/json");
-
-    let raw = JSON.stringify({
-      username: userInfo.email,
-      password: userInfo.password,
-    });
-
-    fetch("http://localhost:4000/auth/signin", {
-      method: "POST",
-      headers: myHeaders,
-      body: raw,
-      redirect: "follow",
-    })
-      .then((response) => response.text())
-      .then((result) => storeUserInfo(userInfo, result))
-      .then(() => console.log("<= SIGNIN"))
-      .then(() => userSignIn())
-      .catch((error) => console.log("error", error));
-  };
-
-  const signUp = () => {
-    let myHeaders = new Headers();
-    myHeaders.append("Content-Type", "application/json");
-
-    let raw = JSON.stringify({
-      email: userInfo.email,
-      password: userInfo.password,
-      username: userInfo.username,
-    });
-
-    fetch("http://localhost:4000/auth/signup", {
-      method: "POST",
-      headers: myHeaders,
-      body: raw,
-      redirect: "follow",
-    })
-      .then((response) => response.text())
-      .then((result) => storeUserInfo(userInfo, result))
-      .then(() => console.log("<= SIGNUP"))
-      .then(() => userSignIn())
-      .catch((error) => console.log("error", error));
+    if (GUserInputsRefs.username.current?.value)
+      userInfo.username = GUserInputsRefs.username.current.value;
+    userInfo.email = GUserInputsRefs!.email!.current!.value;
+    userInfo.password = GUserInputsRefs!.password!.current!.value;
+    if (userInfo.username && userInfo.email && userInfo.password)
+      signUp(userInfo, userSignIn);
+    else signIn(userInfo, userSignIn);
   };
 
   return (
@@ -143,7 +51,7 @@ export default function Auth() {
           <Form.Group className="mb-3">
             <Form.Label className="form-test ">EMAIL</Form.Label>
             <Form.Control
-              ref={userInputsRefs.email}
+              ref={GUserInputsRefs.email}
               type="email"
               placeholder="name@example.com"
             />
@@ -154,7 +62,7 @@ export default function Auth() {
               PASSWORD
             </Form.Label>
             <Form.Control
-              ref={userInputsRefs.password}
+              ref={GUserInputsRefs.password}
               type="password"
               id="inputPassword5"
               aria-describedby="passwordHelpBlock"
